+++ src/data_for_example/Baiedeschaleurs.csv
@@ -1,0 +1,44 @@
+-64.6161111111111,48.14
+-64.8172222222222,48.0561111111111
+-64.7672222222222,48.0691666666667
+-64.7958333333333,48.0511111111111
+-65.2352777777778,47.9375
+-64.7208333333333,48.1358333333333
+-64.8613888888889,48.0497222222222
+-64.8716666666667,48.035
+-64.8722222222222,48.0361111111111
+-65.0469444444444,47.9927777777778
+-65.1483333333333,47.9513888888889
+-64.7430555555556,48.0966666666667
+-65.1147222222222,47.9688888888889
+-65.0388888888889,48.0041666666667
+-65.0394444444444,48.0052777777778
+-65.0288888888889,48.0194444444444
+-64.9947222222222,48.0261111111111
+-64.9630555555556,48.0252777777778
+-64.9630555555556,48.0258333333333
+-64.9227777777778,48.065
+-64.9227777777778,48.065
+-64.8208333333333,48.0627777777778
+-65.1225,47.9722222222222
+-65.0116666666667,48.025
+-65.0022222222222,48.045
+-65.2394444444444,47.9388888888889
+-65.1222222222222,47.9819444444444
+-65.1597222222222,47.9519444444445
+-65.1153,47.9723
+-64.8237,48.1017
+-65.2333,47.9476
+-65.2357,47.9335
+-64.6698,48.1358
+-65.1586,47.9683
+-64.8917,48.0665
+-64.8693,48.0475
+-64.8991,48.0582
+-65.0722,48.0
+-65.0662,47.9913
+-65.1018,47.9752
+-65.0489,48.0189
+-64.6534,48.1722
+-64.9395,48.0692
+-64.9628,48.0145
